+++ gulpfile.js
@@ -12,13 +12,11 @@ var sassLint    = require('gulp-sass-lint');
 var sourcemaps  = require('gulp-sourcemaps');
 var uglify      = require('gulp-uglify');
 
-var localServer = "http://localhost/imhh-fs/public";
-var sourcePath  = "gulpfiles/";
-var publicPath  = "public/";
+var config      = require('./gulpfile-config.json');
 
 // processing scss to css and minify result
 function scss() {
-    return gulp.src(sourcePath + 'scss/styles.scss')
+    return gulp.src(config.sourcePath + 'scss/styles.scss')
         .pipe(sourcemaps.init())
         .pipe(sass())
         .pipe(prefixer({
@@ -27,15 +25,15 @@ function scss() {
         }))
         .pipe(minifyCss({compatibility: 'ie8'}))
         .pipe(sourcemaps.write('./'))
-        .pipe(gulp.dest(publicPath + 'css/'));
+        .pipe(gulp.dest(config.publicPath + 'css/'));
 }
 
 // lint scss files
 function scssLint() {
     return gulp.src([
-            sourcePath + 'scss/**/*.scss',
+            config.sourcePath + 'scss/**/*.scss',
             // exclude third party and special files
-            '!' + sourcePath + 'scss/module/_datatables.scss'
+            '!' + config.sourcePath + 'scss/module/_datatables.scss'
         ])
         .pipe(sassLint(require('./scss-lint.json')))
         .pipe(sassLint.format())
@@ -48,26 +46,26 @@ function js() {
             'node_modules/jquery/dist/jquery.js',
             'node_modules/bootstrap/dist/js/bootstrap.bundle.js',
             'node_modules/@fortawesome/fontawesome-free/js/all.js',
-            sourcePath + 'js/lib/**/*.js',
+            config.sourcePath + 'js/lib/**/*.js',
             'node_modules/cookieconsent/src/cookieconsent.js',
             'node_modules/cssuseragent/cssua.js',
             'node_modules/datatables/media/js/jquery.dataTables.js',
             'node_modules/datatables.net-bs4/js/dataTables.bootstrap4.js',
-            sourcePath + 'js/plugin/**/*.js',
-            sourcePath + 'js/module/**/*.js',
-            sourcePath + 'js/scripts.js'
+            config.sourcePath + 'js/plugin/**/*.js',
+            config.sourcePath + 'js/module/**/*.js',
+            config.sourcePath + 'js/scripts.js'
         ])
         .pipe(sourcemaps.init())
         .pipe(concat('scripts.js'))
         .pipe(uglify())
         .pipe(sourcemaps.write('./'))
-        .pipe(gulp.dest(publicPath + 'js/'));
+        .pipe(gulp.dest(config.publicPath + 'js/'));
 }
 
 // lint js files
 function jsLint() {
     return gulp.src([
-            sourcePath + 'js/**/*.js'
+            config.sourcePath + 'js/**/*.js'
         ])
         .pipe(eslint(require('./js-lint.json')))
         .pipe(eslint.format())
@@ -77,15 +75,15 @@ function jsLint() {
 // copy all json files and minify
 function json() {
     return gulp.src([
-            sourcePath + 'json/**/*.json'
+            config.sourcePath + 'json/**/*.json'
         ])
         .pipe(minifyJson())
-        .pipe(gulp.dest(publicPath + 'json/'));
+        .pipe(gulp.dest(config.publicPath + 'json/'));
 }
 
 // compress images
 function img() {
-    return gulp.src(sourcePath + 'img/**/*.{png,gif,jpg,jpeg,ico,xml,json,svg}')
+    return gulp.src(config.sourcePath + 'img/**/*.{png,gif,jpg,jpeg,ico,xml,json,svg}')
         .pipe(imagemin([
             imagemin.gifsicle({interlaced: true}),
             imagemin.jpegtran({progressive: true}),
@@ -97,16 +95,16 @@ function img() {
                 ]
             })
         ]))
-        .pipe(gulp.dest(publicPath + 'img/'));
+        .pipe(gulp.dest(config.publicPath + 'img/'));
 }
 
 // copy font files
 function font() {
     return gulp.src([
             'node_modules/@fortawesome/fontawesome-free/webfonts/**',
-            sourcePath + 'font/**'
+            config.sourcePath + 'font/**'
         ])
-        .pipe(gulp.dest(publicPath + 'font/'));
+        .pipe(gulp.dest(config.publicPath + 'font/'));
 }
 
 // compress and copy svg files
@@ -114,7 +112,7 @@ function svg() {
     return gulp.src([
 //            'node_modules/@fortawesome/fontawesome-free/svgs/**',
 //            'node_modules/@fortawesome/fontawesome-free/sprites/**',
-            sourcePath + 'svg/**/*.svg'
+            config.sourcePath + 'svg/**/*.svg'
         ])
         .pipe(imagemin([
             imagemin.svgo({
@@ -124,18 +122,18 @@ function svg() {
                 ]
             })
         ]))
-        .pipe(gulp.dest(publicPath + 'svg/'));
+        .pipe(gulp.dest(config.publicPath + 'svg/'));
 }
 
 // clean up folders
 function cleanUp() {
     return del([
-            publicPath + 'css/**/*',
-            publicPath + 'js/**/*',
-            publicPath + 'img/**/*',
-            publicPath + 'json/**/*',
-            publicPath + 'font/**/*',
-            publicPath + 'svg/**/*'
+            config.publicPath + 'css/**/*',
+            config.publicPath + 'js/**/*',
+            config.publicPath + 'img/**/*',
+            config.publicPath + 'json/**/*',
+            config.publicPath + 'font/**/*',
+            config.publicPath + 'svg/**/*'
         ]);
 }
 
@@ -143,7 +141,7 @@ function cleanUp() {
 function browserSyncInit(done) {
     // start browsersync
     browserSync.init({
-        proxy: localServer
+        proxy: config.localServer
     });
     done();
 }
@@ -157,35 +155,35 @@ function browserSyncReload(done) {
 // watch files
 function watch() {
     // watch scss files
-    gulp.watch(sourcePath + 'scss/**', gulp.series(scss, scssLint));
+    gulp.watch(config.sourcePath + 'scss/**', gulp.series(scss, scssLint));
     // watch js files
-    gulp.watch(sourcePath + 'js/**', gulp.series(js, jsLint));
+    gulp.watch(config.sourcePath + 'js/**', gulp.series(js, jsLint));
     // watch images
-    gulp.watch(sourcePath + 'img/**', img);
+    gulp.watch(config.sourcePath + 'img/**', img);
     // watch json files
-    gulp.watch(sourcePath + 'json/**', json);
+    gulp.watch(config.sourcePath + 'json/**', json);
     // watch fonts
-    gulp.watch(sourcePath + 'font/**', font);
+    gulp.watch(config.sourcePath + 'font/**', font);
     // watch svg
-    gulp.watch(sourcePath + 'svg/**', svg);
+    gulp.watch(config.sourcePath + 'svg/**', svg);
 }
 
 // watch files and reload browser on file change
-function watchFiles() {
+function watchAndReload() {
     // watch scss files
-    gulp.watch(sourcePath + 'scss/**', gulp.series(scss, scssLint));
+    gulp.watch(config.sourcePath + 'scss/**', gulp.series(scss, scssLint));
     // watch js files
-    gulp.watch(sourcePath + 'js/**', gulp.series(js, jsLint));
+    gulp.watch(config.sourcePath + 'js/**', gulp.series(js, jsLint));
     // watch images
-    gulp.watch(sourcePath + 'img/**', img);
+    gulp.watch(config.sourcePath + 'img/**', img);
     // watch json files
-    gulp.watch(sourcePath + 'json/**', json);
+    gulp.watch(config.sourcePath + 'json/**', json);
     // watch fonts
-    gulp.watch(sourcePath + 'font/**', font);
+    gulp.watch(config.sourcePath + 'font/**', font);
     // watch svg
-    gulp.watch(sourcePath + 'svg/**', svg);
+    gulp.watch(config.sourcePath + 'svg/**', svg);
     
-    gulp.watch(publicPath + '**/*.{css,eot,ico,js,json,jpg,otf,png,svg,ttf,woff,woff2}', browserSyncReload);
+    gulp.watch(config.publicPath + '**/*.{css,eot,ico,js,json,jpg,otf,png,svg,ttf,woff,woff2}', browserSyncReload);
     gulp.watch('{templates,locale,config,src}/**/*.{php,twig}', browserSyncReload);
 }
 
@@ -199,7 +197,7 @@ exports.font = font;
 exports.svg = svg;
 exports.cleanUp = cleanUp;
 exports.watch = watch;
-exports.watchFiles = watchFiles;
+exports.watchAndReload = watchAndReload;
 exports.browserSyncInit = browserSyncInit;
 exports.browserSyncReload = browserSyncReload;
 
@@ -207,4 +205,4 @@ exports.browserSyncReload = browserSyncReload;
 gulp.task('build', gulp.series(cleanUp, scss, js, scssLint, jsLint, json, img, font, svg));
 
 // default task if just called gulp
-gulp.task('default', gulp.parallel(watchFiles, browserSyncInit));
+gulp.task('default', gulp.parallel(watchAndReload, browserSyncInit));