+++ gulpfiles/js/module/langswitch.js
@@ -1,19 +1,6 @@
 'use strict';
 
 /**
- * @param {string} cname
- * @param {string} cvalue
- * @param {integer} exdays
- * @returns {undefined}
- */
-function setCookie(cname, cvalue, exdays) {
-    var d = new Date();
-    d.setTime(d.getTime() + (exdays * 24 * 60 * 60 * 1000));
-    var expires = 'expires='+d.toUTCString();
-    document.cookie = cname + '=' + cvalue + ';' + expires + ';path=/';
-}
-
-/**
  * @param {object} $
  * @returns {undefined}
  */